+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 2.4.13 (Unreleased)
 - [Enhancement] Report ignored repartitioning to lower number of partitions in declarative topics.
 - [Fix] Fix `Karafka::Errors::MissingBootFileError` when debugging in VScode with ruby-lsp.
+- [Fix] Validate swarm cli and always parse options from argv (roelbondoc)
 
 ## 2.4.12 (2024-09-17)
 - **[Feature]** Provide Adaptive Iterator feature as a fast alternative to Long-Running Jobs (Pro).