+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Karafka Framework Changelog
 
 ## 2.4.13 (Unreleased)
-- [Enhancement] Make declarative topics return different exit codes on migrable/non-mibrable states (0 - no changes, 2 - changes).
+- [Enhancement] Make declarative topics return different exit codes on migrable/non-migrable states (0 - no changes, 2 - changes).
 - [Enhancement] Report ignored repartitioning to lower number of partitions in declarative topics.
 - [Fix] Fix `Karafka::Errors::MissingBootFileError` when debugging in VScode with ruby-lsp.
 - [Fix] Require `karafka-core` `>=` `2.4.4` to prevent dependencies conflicts.